scripts/verification/requirements.txt
requests>=2.31
pyyaml>=6.0
beautifulsoup4>=4.12
lxml>=4.9
feedparser>=6.0
pdfplumber>=0.10
